Resolved Multiple Issues w/ Navbar, Footer, HomePage
Made it so that the navbar and footer will appear on the entire HTML element. Need to figure out how to route the Navbar to new links. I also resolved the navbar mobile view for now. Need some refinements as to where the text is being displayed, etc. I also need to revise the component structure as well, each non-routing component needs to have a page.tsx file in order to be able to properly routed. Unopinionated convention of next.js.