Currently trying to figure out how to get the requestData from the json body to the supabase/stripe.
I learned that changing the NextResponse functions to POST instead of GET worked, though I don't necessarily know how this worked or the repercussions of it.
Need to hook it up to stripe so that the user_profiles column gets updated. The current supabase update function isn't working right now so that is the next step.