Established connection between supabase database/stripe webhook
Disabled RLS for the products/prices & ran the webhook locally through the stripe CLI instead of using the endpoints with the host. In the future, I will use the host endpoint.
Prices & products now write to the database & display to the database in the Pricing Plans Page.
Need to establish the subscriptions table of the Database. Also need to work on having it say that the user is subscribed and how to have subscription based content